+++ docs/changelog.rst
@@ -19,6 +19,7 @@ Changelog
 * :bug:`-` Depositing native token + wrapped native token in Aaave v3 in the same transaction will now be properly decoded.
 * :feature:`11702` Zerox Base Swaps through the latest settler  will now be properly decoded.
 * :bug:`-` ZKSynclite batch withdrawal on the ethereum side should now be properly decoded.
+* :bug:`-` Claiming AAVE rewards and immediately restaking them will now be properly decoded.
 
 * :release:`1.42.0 <2025-02-20>`
 * :feature:`-` Octant v2 migration is now supported and GLM locked in v2 detected.